integration des tests

+++ src/main/java/org/project/service/UserService.java
@@ -2,6 +2,8 @@ package org.project.service;
 
 import org.project.entite.User;
 import org.project.repositorie.UserRepository;
+import org.project.Enum.TypeRole;
+import org.project.Util.PasswordUtil;
 
 import java.util.List;
 
@@ -17,16 +19,46 @@ public class UserService {
     }
 
     public void createUser(User user) {
-        userRepository.createUser(user);
+        if (user != null && user.getEmail() != null && !user.getEmail().isEmpty()) {
+            String hashedPassword = PasswordUtil.hashPassword(user.getMot_de_pass());
+            user.setMot_de_pass(hashedPassword);
+            userRepository.createUser(user);
+        } else {
+            throw new IllegalArgumentException("L'utilisateur ou l'email est invalide");
+        }
     }
+
     public void deleteUser(Long id) {
-        userRepository.deleteUser(id);
+        User user = userRepository.getUserById(id.intValue());
+        if (user != null) {
+            userRepository.deleteUser(id);
+        } else {
+            throw new IllegalArgumentException("Utilisateur introuvable");
+        }
     }
 
     public void updateUser(User user) {
-        userRepository.updateUser(user);
+        if (user != null ) {
+            User existingUser = userRepository.getUserById(user.getId());
+            if (existingUser != null) {
+                existingUser.setName(user.getName());
+                existingUser.setEmail(user.getEmail());
+
+                if (user.getMot_de_pass() != null && !user.getMot_de_pass().isEmpty()) {
+                    existingUser.setMot_de_pass(PasswordUtil.hashPassword(user.getMot_de_pass()));
+                }
+
+                existingUser.setRole(user.getRole());
+                userRepository.updateUser(existingUser);
+            } else {
+                throw new IllegalArgumentException("Utilisateur introuvable");
+            }
+        } else {
+            throw new IllegalArgumentException("Données utilisateur invalides ou identifiant manquant");
+        }
     }
-    public User getUserById(int id){
-      return   userRepository.getUserById(id);
+
+    public User getUserById(int id) {
+        return userRepository.getUserById(id);
     }
 }
